# graphqlbackend: Reduce allocations in zoektIndexedRepos