--- a/cmd/frontend/graphqlbackend/textsearch_test.go
+++ b/cmd/frontend/graphqlbackend/textsearch_test.go
@@ -393,20 +393,20 @@ func Test_zoektSearchHEAD(t *testing.T) {
 	zeroTimeoutCtx, cancel := context.WithTimeout(context.Background(), 0)
 	defer cancel()
 	type args struct {
-		ctx              context.Context
-		query            *search.PatternInfo
-		indexedRevisions map[*search.RepositoryRevisions]string
-		repos            []*search.RepositoryRevisions
-		useFullDeadline  bool
-		searcher         zoekt.Searcher
-		opts             zoekt.SearchOptions
-		since            func(time.Time) time.Duration
+		ctx             context.Context
+		query           *search.PatternInfo
+		indexedCommits  map[*search.RepositoryRevisions]string
+		repos           []*search.RepositoryRevisions
+		useFullDeadline bool
+		searcher        zoekt.Searcher
+		opts            zoekt.SearchOptions
+		since           func(time.Time) time.Duration
 	}
 
 	singleRepositoryRevisions := []*search.RepositoryRevisions{
 		{Repo: &types.Repo{}},
 	}
-	singleIndexedRevisions := map[*search.RepositoryRevisions]string{
+	singleIndexedCommits := map[*search.RepositoryRevisions]string{
 		singleRepositoryRevisions[0]: "abc",
 	}
 
@@ -421,14 +421,14 @@ func Test_zoektSearchHEAD(t *testing.T) {
 		{
 			name: "returns no error if search completed with no matches before timeout",
 			args: args{
-				ctx:              context.Background(),
-				query:            &search.PatternInfo{PathPatternsAreRegExps: true},
-				indexedRevisions: singleIndexedRevisions,
-				repos:            singleRepositoryRevisions,
-				useFullDeadline:  false,
-				searcher:         &fakeSearcher{result: &zoekt.SearchResult{}},
-				opts:             zoekt.SearchOptions{MaxWallTime: time.Second},
-				since:            func(time.Time) time.Duration { return time.Second - time.Millisecond },
+				ctx:             context.Background(),
+				query:           &search.PatternInfo{PathPatternsAreRegExps: true},
+				indexedCommits:  singleIndexedCommits,
+				repos:           singleRepositoryRevisions,
+				useFullDeadline: false,
+				searcher:        &fakeSearcher{result: &zoekt.SearchResult{}},
+				opts:            zoekt.SearchOptions{MaxWallTime: time.Second},
+				since:           func(time.Time) time.Duration { return time.Second - time.Millisecond },
 			},
 			wantFm:            nil,
 			wantLimitHit:      false,
@@ -438,14 +438,14 @@ func Test_zoektSearchHEAD(t *testing.T) {
 		{
 			name: "returns error if max wall time is exceeded but no matches have been found yet",
 			args: args{
-				ctx:              context.Background(),
-				query:            &search.PatternInfo{PathPatternsAreRegExps: true},
-				indexedRevisions: singleIndexedRevisions,
-				repos:            singleRepositoryRevisions,
-				useFullDeadline:  false,
-				searcher:         &fakeSearcher{result: &zoekt.SearchResult{}},
-				opts:             zoekt.SearchOptions{MaxWallTime: time.Second},
-				since:            func(time.Time) time.Duration { return time.Second },
+				ctx:             context.Background(),
+				query:           &search.PatternInfo{PathPatternsAreRegExps: true},
+				indexedCommits:  singleIndexedCommits,
+				repos:           singleRepositoryRevisions,
+				useFullDeadline: false,
+				searcher:        &fakeSearcher{result: &zoekt.SearchResult{}},
+				opts:            zoekt.SearchOptions{MaxWallTime: time.Second},
+				since:           func(time.Time) time.Duration { return time.Second },
 			},
 			wantFm:            nil,
 			wantLimitHit:      false,
@@ -455,14 +455,14 @@ func Test_zoektSearchHEAD(t *testing.T) {
 		{
 			name: "returns error if context timeout already passed",
 			args: args{
-				ctx:              zeroTimeoutCtx,
-				query:            &search.PatternInfo{PathPatternsAreRegExps: true},
-				indexedRevisions: singleIndexedRevisions,
-				repos:            singleRepositoryRevisions,
-				useFullDeadline:  true,
-				searcher:         &fakeSearcher{result: &zoekt.SearchResult{}},
-				opts:             zoekt.SearchOptions{},
-				since:            func(time.Time) time.Duration { return 0 },
+				ctx:             zeroTimeoutCtx,
+				query:           &search.PatternInfo{PathPatternsAreRegExps: true},
+				indexedCommits:  singleIndexedCommits,
+				repos:           singleRepositoryRevisions,
+				useFullDeadline: true,
+				searcher:        &fakeSearcher{result: &zoekt.SearchResult{}},
+				opts:            zoekt.SearchOptions{},
+				since:           func(time.Time) time.Duration { return 0 },
 			},
 			wantFm:            nil,
 			wantLimitHit:      false,
@@ -472,14 +472,14 @@ func Test_zoektSearchHEAD(t *testing.T) {
 		{
 			name: "returns error if searcher returns an error",
 			args: args{
-				ctx:              context.Background(),
-				query:            &search.PatternInfo{PathPatternsAreRegExps: true},
-				indexedRevisions: singleIndexedRevisions,
-				repos:            singleRepositoryRevisions,
-				useFullDeadline:  true,
-				searcher:         &errorSearcher{err: errors.New("womp womp")},
-				opts:             zoekt.SearchOptions{},
-				since:            func(time.Time) time.Duration { return 0 },
+				ctx:             context.Background(),
+				query:           &search.PatternInfo{PathPatternsAreRegExps: true},
+				indexedCommits:  singleIndexedCommits,
+				repos:           singleRepositoryRevisions,
+				useFullDeadline: true,
+				searcher:        &errorSearcher{err: errors.New("womp womp")},
+				opts:            zoekt.SearchOptions{},
+				since:           func(time.Time) time.Duration { return 0 },
 			},
 			wantFm:            nil,
 			wantLimitHit:      false,
@@ -489,7 +489,7 @@ func Test_zoektSearchHEAD(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			gotFm, gotLimitHit, gotReposLimitHit, err := zoektSearchHEAD(tt.args.ctx, tt.args.query, tt.args.repos, tt.args.indexedRevisions, tt.args.useFullDeadline, tt.args.searcher, tt.args.opts, tt.args.since)
+			gotFm, gotLimitHit, gotReposLimitHit, err := zoektSearchHEAD(tt.args.ctx, tt.args.query, tt.args.repos, tt.args.indexedCommits, tt.args.useFullDeadline, tt.args.searcher, tt.args.opts, tt.args.since)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("zoektSearchHEAD() error = %v, wantErr = %v", err, tt.wantErr)
 				return
@@ -666,7 +666,7 @@ func Test_zoektIndexedRepos(t *testing.T) {
 	zoekt := &searchbackend.Zoekt{Client: &fakeSearcher{repos: zoektRepoList}}
 	ctx := context.Background()
 
-	indexed, unindexed, indexedRevisions, err := zoektIndexedRepos(ctx, zoekt, repos)
+	indexed, unindexed, indexedCommits, err := zoektIndexedRepos(ctx, zoekt, repos)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -678,10 +678,6 @@ func Test_zoektIndexedRepos(t *testing.T) {
 		{"indexed", indexed, makeRepositoryRevisions("foo/indexed-one@", "foo/indexed-two@", "foo/indexed-three-no-HEAD@")},
 		{"unindexed", unindexed, makeRepositoryRevisions("foo/unindexed-one", "foo/unindexed-two")},
 	} {
-		if len(assertion.have) != len(assertion.want) {
-			t.Fatalf("%s has wrong length: %d", assertion.name, len(assertion.have))
-		}
-
 		sort.Slice(assertion.have, sortRepoRevsByName(assertion.have))
 		sort.Slice(assertion.want, sortRepoRevsByName(assertion.want))
 
@@ -691,14 +687,14 @@ func Test_zoektIndexedRepos(t *testing.T) {
 		}
 	}
 
-	wantIndexedRevisions := map[*search.RepositoryRevisions]string{
+	wantIndexedCommits := map[*search.RepositoryRevisions]string{
 		repos[0]: "deadbeef",
 		repos[1]: "deadbeef",
 	}
 
-	if !reflect.DeepEqual(indexedRevisions, wantIndexedRevisions) {
-		diff := cmp.Diff(indexedRevisions, wantIndexedRevisions)
-		t.Fatalf("indexedRevisions has wrong revisions. diff=%s", diff)
+	if !reflect.DeepEqual(indexedCommits, wantIndexedCommits) {
+		diff := cmp.Diff(indexedCommits, wantIndexedCommits)
+		t.Fatalf("indexedCommits has wrong revisions. diff=%s", diff)
 	}
 }