### Original delta
```
name                   old time/op    new time/op    delta
SearchResults-16         6.13ms ±22%    6.04ms ±19%     ~     (p=0.912 n=10+10)
_zoektIndexedRepos-16    1.83ms ± 4%    1.30ms ± 2%  -29.02%  (p=0.000 n=9+9)

name                   old alloc/op   new alloc/op   delta
SearchResults-16         2.32MB ± 0%    2.39MB ± 0%   +2.84%  (p=0.000 n=10+10)
_zoektIndexedRepos-16    2.50MB ± 0%    1.16MB ± 0%  -53.82%  (p=0.000 n=10+10)

name                   old allocs/op  new allocs/op  delta
SearchResults-16          20.7k ± 0%     15.8k ± 0%  -23.57%  (p=0.000 n=8+10)
_zoektIndexedRepos-16     20.1k ± 0%      0.0k ± 0%  -99.96%  (p=0.000 n=10+10)
```

### Updated delta

```
name                   old time/op    new time/op    delta
SearchResults-16         5.37ms ±13%    4.79ms ±11%  -10.73%  (p=0.004 n=10+10)
_zoektIndexedRepos-16    2.41ms ± 3%    0.58ms ± 4%  -76.02%  (p=0.000 n=10+10)

name                   old alloc/op   new alloc/op   delta
SearchResults-16         2.32MB ± 0%    1.97MB ± 0%  -14.88%  (p=0.000 n=10+10)
_zoektIndexedRepos-16    2.49MB ± 0%    0.17MB ± 1%  -93.36%  (p=0.000 n=10+10)

name                   old allocs/op  new allocs/op  delta
SearchResults-16          20.7k ± 0%     15.6k ± 0%  -24.28%  (p=0.000 n=8+10)
_zoektIndexedRepos-16     20.1k ± 0%      0.0k ± 0%  -99.97%  (p=0.000 n=10+10)

```