@tsenart 2019-07-01
@mrnugget, @keegancsmith: I addressed all review comments. Please review again! Things got even faster. (Check PR description).