@tsenart 2019-07-01
And here's the delta for the more recently added `BenchmarkIntegrationSearchResults`.

```
name                         old time/op    new time/op    delta
IntegrationSearchResults-16    23.5ms ±13%    22.7ms ±11%    ~     (p=0.138 n=19+19)

name                         old alloc/op   new alloc/op   delta
IntegrationSearchResults-16    8.11MB ± 0%    7.75MB ± 0%  -4.36%  (p=0.000 n=18+18)

name                         old allocs/op  new allocs/op  delta
IntegrationSearchResults-16      166k ± 0%      161k ± 0%  -3.02%  (p=0.000 n=15+16)
```