@keegancsmith 2019-07-01
> And here's the delta for the more recently added `BenchmarkIntegrationSearchResults`.

Surprisingly small difference. Is it possible you mixed up the base you compared against? Or is it due to not using caching in the benchmark? We should use caching in the benchmark since it doesn't affect coverage. Just make sure you close the client so we shutdown those goroutines.